fix random errors in pages to isolate errors with auth. add next-auth to project. add error handling
Added error handling for 404s that doesn't reload constantly in dev. Added next-auth to project and set up initial provider (github). Resolved some pre-render errors that could be masking issues with next-auth.
