+++ _projects/german_trainer.md
