+++ app/[...not-found]/page.tsx
@@ -1,0 +1,5 @@
+import { notFound } from 'next/navigation';
+
+const NotFoundCatchAll = () => notFound();
+
+export default NotFoundCatchAll;
