+++ app/editor/page.tsx
@@ -1,240 +1,15 @@
-'use client';
+import dynamic from 'next/dynamic';
 
-import { useCallback, useEffect, useRef, useState } from 'react';
-import Code from '@editorjs/code';
-import { EditorHeader } from './editorHeader';
-import EditorJS from '@editorjs/editorjs';
-import { EditorPostOutput } from '@customTypes/editorTypes';
-import Header from '@editorjs/header';
-import Image from '@editorjs/image';
-import List from '@editorjs/list';
-import MermaidTool from 'editorjs-mermaid';
-import { POST_URL } from '@constants/urls';
-import { PostMetaData } from './postMetadata';
-import Quote from '@editorjs/quote';
+const EditorNoSSR = dynamic(() => import('./editorPage'), { ssr: false });
 
-import styled from 'styled-components';
-import { v4 as uuidv4 } from 'uuid';
-
-const EditorWrapper = styled.section`
-  border: 1px solid #aaaaaa;
-`;
-
-const InputLabel = styled.label`
-  display: block;
-`;
-
-const TitleInput = styled.input`
-  font-family: serif;
-  margin-bottom: 0.5rem;
-  padding: 0.25rem;
-`;
-
-const Editor = () => {
-  const [modifiedPost, setModifiedPost] = useState<
-    EditorPostOutput | undefined
-  >();
-  const [editor, setEditor] = useState<EditorJS | undefined>(undefined);
-  const [savable, setSavable] = useState(false);
-  const editorRef = useRef(null);
-  const titleRef = useRef<HTMLInputElement>(null);
-
-  const isFormSavable = useCallback(() => {
-    if (!editorRef.current) {
-      console.warn('no current editor ref');
-      setSavable(false);
-    } else {
-      const titleReady = (() => {
-        if (
-          titleRef.current?.value?.length &&
-          titleRef.current?.value?.length > 0
-        ) {
-          return true;
-        }
-        return false;
-      })();
-
-      const dirty = isDirty(editorRef.current);
-
-      setSavable(titleReady && dirty);
-    }
-  }, []);
-
-  // on load to set up editor object
-  useEffect(() => {
-    const neweditor: EditorJS = new EditorJS({
-      /**
-       * Id of Element that should contain the Editor
-       */
-      holder: 'editorjs',
-      onChange: () => {
-        isFormSavable();
-      },
-      tools: {
-        code: Code,
-        header: Header,
-        list: List,
-        quote: Quote,
-        mermaidTool: MermaidTool,
-        image: Image,
-      },
-    });
-
-    neweditor.isReady
-      .then(() => {
-        setEditor(neweditor);
-      })
-      .catch((reason) => {
-        console.warn(`Editor.js initialization failed because of ${reason}`);
-      });
-
-    return () => {
-      neweditor.destroy();
-    };
-  }, [isFormSavable]);
-
-  const isDirty = (el: HTMLDivElement) => {
-    const editor = el.getElementsByClassName(
-      'codex-editor__redactor'
-    )[0] as HTMLDivElement;
-
-    if (editor && editor.children) {
-      const state = Array.from(editor.children).some((el) => {
-        const childrenWithText: HTMLElement[] = (
-          Array.from(el.children) as HTMLElement[]
-        ).filter((el) => el.innerText.length > 0);
-        return childrenWithText.length > 0;
-      });
-
-      return state;
-    }
-    return false;
-  };
-
-  const writeToPosts = async (data: Partial<EditorPostOutput>) => {
-    const postData = {
-      ...data,
-      post_id: modifiedPost?.post_id || uuidv4(),
-      time_created: modifiedPost?.time_created || data.time || Date.now(),
-      time_modified: data.time || Date.now(),
-    };
-    console.log({ postData });
-
-    await fetch(POST_URL, {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify(postData),
-    });
-  };
-
-  const newPost = () => {
-    if (editor) {
-      editor.clear();
-    }
-    if (titleRef.current) {
-      titleRef.current.value = '';
-    }
-  };
-
-  const selectPostToUpdate = async (post_id: string) => {
-    const results = await fetch(`${POST_URL}?post_id=${post_id}`, {
-      method: 'GET',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-    });
-
-    const data = await results.json();
-    const postData = data.posts[0];
-
-    if (postData && editor) {
-      const { title, time_created, blocks, post_id } = postData;
-
-      console.log({ postData });
-      setModifiedPost(postData);
-
-      if (titleRef.current) {
-        titleRef.current.value = title;
-      }
-
-      editor.render({ blocks });
-    } else {
-      // handle error
-    }
-  };
-
-  const savePostHandler = (published: boolean) => {
-    if (editor) {
-      editor
-        .save()
-        .then(async (outputData) => {
-          const data = {
-            ...outputData,
-            published,
-            title: titleRef.current?.value || 'default title',
-          };
-          await writeToPosts(data);
-          newPost();
-        })
-        .catch((error) => {
-          console.warn('Saving failed: ', error);
-        });
-    }
-  };
-
+const EditorPage = () => {
   return (
-    <main>
-      <EditorHeader selectPostToUpdate={selectPostToUpdate} />
-      <InputLabel htmlFor="title-input">Title</InputLabel>
-      <TitleInput
-        type="text"
-        id="title-input"
-        maxLength={255}
-        ref={titleRef}
-        required
-        onChange={isFormSavable}
-      />
-      {modifiedPost && PostMetaData(modifiedPost)}
-      <EditorWrapper id="editorjs" ref={editorRef}></EditorWrapper>
-      {savable &&
-        ((post?: EditorPostOutput) => {
-          const { post_id: postId = '', published = false } = post || {};
-
-          return published ? (
-            ''
-          ) : (
-            <button
-              onClick={() => {
-                savePostHandler(false);
-              }}
-            >
-              {postId ? 'update' : 'save'} draft
-            </button>
-          );
-        })(modifiedPost)}
-
-      {savable && (
-        <button
-          onClick={() => {
-            savePostHandler(true);
-          }}
-        >
-          {modifiedPost ? 'update' : 'save'} post
-        </button>
-      )}
-      {modifiedPost && (
-        <button
-          onClick={() => {
-            newPost();
-          }}
-        >
-          new post (discard edits)
-        </button>
-      )}
-    </main>
+    <>
+      <EditorNoSSR />
+    </>
   );
 };
 
-export default Editor;
+export default EditorPage;
+
+// CREDIT WHERE CREDIT IS DUE: https://stackoverflow.com/a/73589166
