+++ app/layout.tsx
@@ -4,6 +4,7 @@ import '@scss/style.scss';
 
 import { MainFooter } from '@components/main-footer';
 import { MainHeader } from '@components/main-header';
+import { SessionProvider } from 'next-auth/react';
 import StyledComponentsRegistry from '@utilities/registry';
 
 export default function RootLayout({
@@ -14,11 +15,13 @@ export default function RootLayout({
   return (
     <html lang="en">
       <StyledComponentsRegistry>
-        <body>
-          <MainHeader />
-          {children}
-          <MainFooter />
-        </body>
+        <SessionProvider>
+          <body>
+            <MainHeader />
+            {children}
+            <MainFooter />
+          </body>
+        </SessionProvider>
       </StyledComponentsRegistry>
     </html>
   );
