+++ package.json
@@ -23,7 +23,8 @@
     "editorjs-mermaid": "^1.0.0",
     "glob": "^10.3.1",
     "gray-matter": "^4.0.3",
-    "next": "^13.4.0",
+    "next": "^13.4.19",
+    "next-auth": "^4.23.1",
     "pg": "^8.11.3",
     "raw-loader": "^4.0.2",
     "react": "18.2.0",
